the libraries in lib_LTLIBRARIES must be ordered by dependency
Apparently the libraries listed in lib_LTLIBRARIES must be sorted by
dependency, or an install-time link error occurs.

Ouch!
(You might think this is what "make" is for.  Talk to the hand!
This is a longstanding known but undocumented bug.)

Add script src/scripts/tsort-libs to generate an appropriate order by
conducting a topological sort.